+++ config/initializers/cors.rb
@@ -7,10 +7,11 @@
 
 Rails.application.config.middleware.insert_before 0, Rack::Cors do
   allow do
-    origins "localhost:3001", "https://drone-flight-log.vercel.app"
+    origins "http://localhost:3001", "https://drone-flight-log.vercel.app"
 
     resource "*",
       headers: :any,
-      methods: [:get, :post, :put, :patch, :delete, :options, :head]
-  end
+      methods: [:get, :post, :put, :patch, :delete, :options, :head],
+      credentials: true
+    end
 end
